+++ CHANGELOG.md
@@ -11,6 +11,12 @@
   This includes being able to directly access each top-level menu using their
   access keys (e.g. Alt+F for the File menu).
 
+- Item details now has two new automatic tracking modes prioritise the current
+  or playlist selection over the playing item.
+  [[#1842](https://github.com/reupen/columns_ui/pull/1842)]
+
+  All automatic tracking modes were renamed to make the priority order clearer.
+
 - The following dialogues are now resizeable
   [[#1833](https://github.com/reupen/columns_ui/pull/1833),
   [#1835](https://github.com/reupen/columns_ui/pull/1835)],